fleet_vehicles/config.py
from flask import Flask, request, jsonify
from dotenv import load_dotenv
from flask_sqlalchemy import SQLAlchemy
from flask_marshmallow import Marshmallow
from sqlalchemy.orm import relationship, backref
import os

# Initialization
app = Flask(__name__)

# Database configuration
load_dotenv()
# mysql+pymysql://user:password@host/db
app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql+pymysql://' + os.getenv('DB_USER') + '@' + os.getenv('DB_HOST') + '/' + os.getenv('DB_NAME')
app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
db = SQLAlchemy(app)
ma = Marshmallow(app)

