fleet_vehicles/requirements.txt
click==8.1.3
Flask==2.1.2
flask-marshmallow==0.14.0
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
mysqlclient==2.1.0
packaging==21.3
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
zipp==3.8.0
